StartScreen画面の分離とCSS構造の整理を実施
