--- src/components/Header.tsx
@@ -1,3 +1,0 @@
-export const Header = () => (
-  <header className="text-2xl font-bold px-4 py-2 border-b">RPGセーブエディター</header>
-);
