+++ src/components/StartScreen/Logo.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const Logo: React.FC = () => {
+  return <img src="./assets/title.png" className="start-screen-title" draggable={false} />;
+};
