+++ src/components/StartScreen/StartScreen.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import { Logo } from "./Logo";
+import { DropZone } from "./DropZone";
+import "./start-screen.css";
+
+type StartScreenProps = {
+  onLoad: (json: object) => void;
+};
+
+export const StartScreen: React.FC<StartScreenProps> = ({ onLoad }) => {
+  return (
+    <div className="start-screen-container">
+      <Logo />
+      <DropZone onLoad={onLoad} />
+    </div>
+  );
+};
