+++ src/index.css
@@ -13,31 +13,6 @@ body {
   background-color: #000000;
 }
 
-.start-screen-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 2rem;
-  width: 100%;
-  max-width: 600px;
-  margin: 0 auto;
-  padding: 1rem;
-}
-
-.start-screen-title {
-  max-width: 300px;
-  width: 100%;
-  height: auto;
-  margin-top: 40px;
-}
-
-.start-screen-dropzone {
-  max-width: 360px;
-  width: 100%;
-  height: auto;
-  cursor: pointer;
-}
-
 .save-data-editor-container {
   display: flex;
   gap: 12px;
